+++ src/components/Table/HeadTableRow.vue
@@ -6,31 +6,33 @@
 
 <script setup lang="ts">
 import type { IHeadTableRowProps } from './interface/interface';
+defineOptions({
+  name: 'HeadTableRow'
+});
 
 const props = withDefaults(defineProps<IHeadTableRowProps>(), {
   dataTestid: 'TableRow'
-});
-
-defineOptions({
-  name: 'HeadTableRow'
 });
 </script>
 
 <style scoped>
 .head-table-row {
-  border-left: 1px solid transparent;
-  border-right: 1px solid transparent;
-}
+  &:last-child > * {
+    border-bottom: 0.5px solid var(--border-grey);
+  }
 
-.head-table-row:last-child > * {
-  border-bottom: 1px solid var(--border-grey);
-}
+  &:first-child {
+    & > * {
+      border-top: none;
+    }
 
-.head-table-row:first-child > *:first-child {
-  border-top-left-radius: var(--border-radius);
-}
+    & > *:first-child {
+      border-top-left-radius: var(--border-radius);
+    }
 
-.head-table-row:first-child > *:last-child {
-  border-top-right-radius: var(--border-radius);
+    & > *:last-child {
+      border-top-right-radius: var(--border-radius);
+    }
+  }
 }
 </style>